refactor: refactor table code to a new component
